chore(docs): Improved code consistency and formatting across UI examples
# Project State

## Current Focus
Improved code consistency and formatting across UI examples

## Context
The changes enhance code quality and organization in the UI examples, particularly in system monitoring, log monitoring, and widget tutorials.

## Completed
- [x] Refactored `system_monitor.rs` to use `first()` instead of index-based access for process state
- [x] Improved formatting in `log_monitor.rs` by removing unnecessary `as usize` conversion
- [x] Enhanced centering calculations in `widget_tutorial.rs` by removing redundant `.max(0)` calls
- [x] Updated Cargo.lock with dependency version bumps

## In Progress
- [x] Code quality improvements across UI examples

## Blockers
- None identified

## Next Steps
1. Review additional UI examples for similar formatting improvements
2. Verify all examples maintain expected functionality after changes
---
category: chore
scope: docs
---

+++ examples/_cookbook/log_monitor.rs
@@ -139,7 +139,7 @@ impl Widget for LogMonitor {
         // Header
         let title = " Log Monitor ";
         for (i, c) in title.chars().enumerate().take(area.width as usize - 2) {
-            let idx = (i + 2) as usize;
+            let idx = (i + 2);
             p.cells[idx] = Cell {
                 char: c,
                 fg: self.log_viewer.theme.primary,
